Rescol-backend/requirements.txt
asgiref==3.7.2
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
bcrypt==4.1.3
Beaker==1.13.0
boto3==1.33.1
botocore==1.33.1
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.1
distlib==0.3.7
Django==4.1.11
django-cors-headers==4.2.0
django-jsonfield==1.4.1
djangorestframework==3.14.0
djongo==1.3.6
dnspython==2.4.2
fastkml==0.12
filelock==3.12.2
fiona==1.9.5
geopandas==0.14.1
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
jmespath==1.0.1
numpy==1.25.2
packaging==23.1
pandas==2.1.0
platformdirs==3.10.0
psycopg2-binary==2.9.7
PuLP==2.9.0
pump==0.1.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pygeoif==0.7
pymongo==3.12.3
pyOpenSSL==24.1.0
pyproj==3.6.0
pyshp==2.3.1
python-dateutil==2.8.2
pytz==2023.3.post1
s3transfer==0.8.0
service-identity==24.1.0
shapely==2.0.1
six==1.16.0
sqlparse==0.2.4
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.10.0
tzdata==2023.3
urllib3==2.0.7
virtualenv==20.24.2
zope.interface==6.2
